[lldb] Initialize only one type info provider per swift language runtime
The type info provider pointer is actually part of the cache key when
looking up type infos. Since LLDB was creating a brand new type info
provider on every type info lookup, LLDB was getting cache misses 100%
of the time. Instantiate only one type info provider instead.

rdar://80001304
